pyproject.toml
[project]
name = "mwx"
version = "0.1.0"
description = "QoL tool for MyWallet Android app"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = ["black>=25.1.0", "pytest>=8.4.1", "ruff>=0.12.9"]

[tool.ruff.lint]
ignore = ["F401"]
